final_updatefor8

+++ src/Controllers/EnvironmentController.php
@@ -2,8 +2,10 @@
 
 namespace RachidLaasri\LaravelInstaller\Controllers;
 
+use App\Models\Access\User\User;
 use Exception;
 use GuzzleHttp\Client;
+use GuzzleHttp\Exception\GuzzleException;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
 use Illuminate\Routing\Redirector;
@@ -12,6 +14,7 @@ use Illuminate\Support\Facades\DB;
 use RachidLaasri\LaravelInstaller\Events\EnvironmentSaved;
 use RachidLaasri\LaravelInstaller\Helpers\EnvironmentManager;
 use Validator;
+use Illuminate\Support\Facades\URL;
 
 class EnvironmentController extends Controller
 {
@@ -45,7 +48,12 @@ class EnvironmentController extends Controller
      */
     public function environmentWizard()
     {
+
         $envConfig = $this->EnvironmentManager->getEnvContent();
+
+        if ((isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] !== 'off') or !empty($_SERVER['HTTP_X_FORWARDED_PROTO']) && $_SERVER['HTTP_X_FORWARDED_PROTO'] == 'https') {
+            URL::forceScheme('https');
+        }
 
         return view('vendor.installer.environment-wizard', compact('envConfig'));
     }
@@ -82,54 +90,66 @@ class EnvironmentController extends Controller
      */
     public function saveWizard(Request $request, Redirector $redirect)
     {
-        $flag = true;
-        $status = 'Success';
-        $rules = config('installer.environment.form.rules');
-        $messages = [
-            'environment_custom.required_if' => trans('installer_messages.environment.wizard.form.name_required'),
-        ];
-        $m = '<ul>';
+           try {
 
-        $validator = Validator::make($request->all(), $rules, $messages);
+               $flag = true;
+               $code = '';
+               $status = 'Success';
+               $rules = config('installer.environment.form.rules');
+               $messages = [
+                   'environment_custom.required_if' => trans('installer_messages.environment.wizard.form.name_required'),
+               ];
+               $m = '<ul>';
 
-        if ($validator->fails()) {
-            $status = 'Error';
-            $flag = false;
-            foreach ($validator->errors()->all() as $error) {
-                $m .= '<li>' . $error . '</li>';
-            }
-        }
+               $validator = Validator::make($request->all(), $rules, $messages);
 
-        if (!$this->checkDatabaseConnection($request)) {
-            //   return $redirect->route('LaravelInstaller::environmentWizard')->withInput()->withErrors([
-            //       'database_connection' => trans('installer_messages.environment.wizard.form.db_connection_failed'),
-            //   ]);
-            $status = 'Error';
-            $flag = false;
-            $m .= '<li>' . trans('installer_messages.environment.wizard.form.db_connection_failed') . '</li>';
-        }
-        $valid = $this->checkValid($request);
-        if ($valid) {
-            $this->confirmValid($request);
-        } else {
-            $status = 'Error';
-            $flag = false;
-            $m .= '<li>' . trans('installer_messages.environment.wizard.form.verification_failed') . '</li>';
-        }
+               if ($validator->fails()) {
+                   $status = 'Error';
+                   $flag = false;
+                   foreach ($validator->errors()->all() as $error) {
+                       $m .= '<li>' . $error . '</li>';
+                   }
+                   $code = 'F3';
+               }
 
-        $m .= '</ul>';
+               if (!$this->checkDatabaseConnection($request)) {
+                   $status = 'Error';
+                   $flag = false;
+                   $m .= '<li>' . trans('installer_messages.environment.wizard.form.db_connection_failed') . '</li>';
+                   $code = 'D6';
+               }
+               $valid = $this->checkValid($request);
 
-        {
-            return json_encode(array('status' => $status, 'message' => $m));
-            if ($flag) {
+               $valid = json_decode($valid, true);
 
-                $results = $this->EnvironmentManager->saveFileWizard($request);
-                event(new EnvironmentSaved($request));
-               // return $redirect->route('LaravelInstaller::database')
-                //    ->with(['results' => $results]);
+               if (@$valid['valid']) {
+                   $this->confirmValid($request);
+               } else {
+                   $status = 'Error';
+                   $flag = false;
+                   $m .= '<li>' . trans('installer_messages.environment.wizard.form.verification_failed') . '</li>';
+                   $code = 'C5';
+               }
+               if (@$valid['status'] == 'Error') {
+                   $status = 'Error';
+                   $flag = false;
+                   $m .= $valid['message'];
+               }
+               $m .= '</ul>';
 
-            }
-        }
+
+               if ($flag) {
+                   $m .= '<li>Installation Success, Installer locked to prevent re-installation</li>';
+                   $results = $this->EnvironmentManager->saveFileWizard($request);
+                   event(new EnvironmentSaved($request));
+                   return json_encode(array('status' => $status, 'message' => $m));
+               } else {
+                   return json_encode(array('status' => $status, 'message' => $m));
+               }
+           } catch (Exception $e){
+                   return json_encode(array('status' => 'Error', 'message' => 'Indicating Unsupported PHP Version & '.$e->getMessage()));
+           }
+
     }
 
     /**
@@ -169,16 +189,25 @@ class EnvironmentController extends Controller
         $config = Config::get('version');
         $build = $config['build'];
         $zone = $config['zone'];
-        $client = new Client(["base_uri" => $zone]);
-        $options = [
-            'form_params' => [
-                "c" => "codepost",
-            ]
-        ];
-        $response = $client->post('confirm/' . $build . '/', $options);
-        $stream = $response->getBody();
-        $contents = $stream->getContents();
-        return $contents;
+
+        try {
+            $client = new Client(["base_uri" => $zone]);
+            $options = [
+                'form_params' => [
+                    "c" => $request->input('purchase_code'),
+                    "mail" => $request->input('user_email'),
+                    "u" => url('/'),
+                ]
+            ];
+            $response = $client->post('confirm/' . $build . '/', $options);
+            $stream = $response->getBody();
+            $contents = $stream->getContents();
+
+            return $contents;
+        } catch (GuzzleException $e) {
+            return json_encode(array('status' => 'Error', 'message' => trans('installer_messages.environment.wizard.form.db_communication_failed')));
+        }
+
     }
 
     private function confirmValid(Request $request)
@@ -187,40 +216,54 @@ class EnvironmentController extends Controller
         $config = Config::get('version');
         $build = $config['build'];
         $zone = $config['zone'];
-        $client = new Client(["base_uri" => $zone]);
-        $options = [
-            'form_params' => [
-                "c" => "codepost",
-            ]
-        ];
-        $response = $client->post('confirm/' . $build . '/valid.php', $options);
-        $stream = $response->getBody();
-        $contents = $stream->getContents();
-        //File Fetch Section Close
 
-        $connection = $request->input('database_connection');
-        DB::purge('mysql');
-        Config::set('database.connections.mysql', [
-            'driver' => $connection,
-            'host' => $request->input('database_hostname'),
-            'port' => $request->input('database_port'),
-            'database' => $request->input('database_name'),
-            'username' => $request->input('database_username'),
-            'password' => $request->input('database_password'),
-        ]);
         try {
-            $conn = DB::connection('mysql');
-            if ($conn->getDatabaseName()) {
-                ini_set('memory_limit', '-1');
-                $conn->unprepared($contents);
-                $conn->commit();
-                return true;
+            $client = new Client(["base_uri" => $zone]);
+            $options = [
+                'form_params' => [
+                    "c" => $request->input('purchase_code'),
+                    "mail" => $request->input('user_email'),
+                      "u"=>url('/'),
+                ]
+            ];
+            $response = $client->post('confirm/' . $build . '/valid.php', $options);
+            $stream = $response->getBody();
+            $contents = $stream->getContents();
+        } catch (GuzzleException $e){
+           return array('status' => 'Error', 'message' => trans('installer_messages.environment.wizard.form.db_communication_failed').$e->getMessage());
 
-            } else {
+        }
 
-                return false;
+
+
+        if (isset($contents)) {
+            $connection = $request->input('database_connection');
+            DB::purge('mysql');
+            Config::set('database.connections.mysql', [
+                'driver' => $connection,
+                'host' => $request->input('database_hostname'),
+                'port' => $request->input('database_port'),
+                'database' => $request->input('database_name'),
+                'username' => $request->input('database_username'),
+                'password' => $request->input('database_password'),
+            ]);
+            try {
+                $conn = DB::connection('mysql');
+                if ($conn->getDatabaseName()) {
+                    ini_set('memory_limit', '-1');
+                    $conn->unprepared($contents);
+                    $conn->commit();
+                    return true;
+
+                } else {
+                    return array('status' => 'Error', 'message' => trans('installer_messages.environment.wizard.form.db_connection_failed'));
+
+                }
+            } catch (\Exception $e) {
+                return array('status' => 'Error', 'message' => trans('installer_messages.environment.wizard.form.db_connection_failed'));
+
             }
-        } catch (\Exception $e) {
+        } else {
             return false;
         }
